move all logic into reusable PONSO(tm)
